更新组件，重构代码

+++ imagepicker/src/main/java/com/lcw/library/imagepicker/utils/MediaFileUtil.java
@@ -177,6 +177,6 @@ public class MediaFileUtil {
     //根据mime类型查看文件类型
     public static int getFileTypeForMimeType(String mimeType) {
         Integer value = sMimeTypeMap.get(mimeType);
-        return (value == null ? 0 : value.intValue());
+        return (value == null ? 0 : value);
     }
 }
